Moved coordinate conversion functions to single file
